refac: extract App
